Rename `hp` package to `hpi`
